temp: mac-zsh: migration plugins

+++ bin/sites-migration-plugins.sh
@@ -19,10 +19,18 @@ echo "Running for ${WP_PATH}"
 /var/scripts/link-repos.sh "${WP_PATH}/wp-content"
 
 # link action disabler
-ln -s "/var/scripts/as-disable-default-runner.php" "${WP_PATH}/wp-content/plugins/"
+link="${WP_PATH}/wp-content/plugins/as-disable-default-runner.php"
+if [ -L "${link}" ]; then
+    echo "${link} already symlinked"
+else
+    ln -s "/var/scripts/as-disable-default-runner.php" "${WP_PATH}/wp-content/plugins/"
+fi
 wp --allow-root --path=${WP_PATH} plugin activate "as-disable-default-runner"
 
-# wp --allow-root --path=${WP_PATH} plugin install co-authors-plus --version=3.5.15 --activate
+# cap
+wp --allow-root --path=${WP_PATH} plugin install co-authors-plus --version=3.6.3
+
+# which versions?
 # wp --allow-root --path=${WP_PATH} plugin install redirection --version=5.4.2 --activate
 # wp --allow-root --path=${WP_PATH} plugin install wordpress-seo --version=22.2 --activate
